Added interface file and code to allow the GUISetUp to get methods and input from MainFrame
Added a reference to GUISetUp to MainFrame to allow GUISetUp to get input and events from MainFrame.  Also made MainFrame inherit from ActionListener so it could get the input from the buttons and send it to the GUISetUp.  Added minor variables and updated the MainFrame to allow for inheritance and removed the main method.  Finished the last bit for RunBuckFarm.